requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
DAWG-Python==0.7.2
docopt==0.6.2
filelock==3.8.0
huggingface-hub==0.10.1
idna==3.4
json2html==1.3.0
numpy==1.23.4
packaging==21.3
pullenti==3.23
pullenti-client==0.6.0
pullenti-wrapper==0.9.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
six==1.16.0
tokenizers==0.13.1
torch==1.13.0
tornado==6.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
